Refactor SlidesPage component and remove pitch page
- Consolidated slide management into the SlidesPage component, enhancing user interaction with new navigation features.
- Implemented touch and click event handling for slide transitions, improving mobile usability.
- Removed the deprecated pitch page, streamlining the codebase and focusing on the updated presentation structure.
- Updated imports and component structure for better organization and maintainability.
